Created input actions and created manager class
We now have a manager class which we can use to access singleton instances of a class such as the user input class. The class is also used to enable events.

For the input actions, there now is a movement action and a aiming action. Both will return a Vector2 where the x and the y variables are modified based on wether or not the selected keys in the input actions have been pressed.